src/api/pyproject.toml
[tool.poetry]
name = "djb_api"
version = "0.1.0"
description = "Screenshot upload service"
authors = ["Jelena Dokic <jelena.dpk@gmail.com>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.3.0"
aiofiles = "^22.1.0"
argon2-cffi = "^21.3.0"
fastapi = "^0.88.0"
passlib = "^1.7.4"
psycopg = "^3.1.6"
pyjwt = "^2.6.0"
python-multipart = "^0.0.5"
uvicorn = "^0.20.0"
websockets = "^10.4"

[tool.poetry.dev-dependencies]
coverage = "^6.5.0"
pytest = "^7.2.0"
pytest-env = "^0.8.1"
python-dotenv = "^0.21.0"
requests = "^2.28.1"

[tool.poetry.group.dev.dependencies]
types-passlib = "^1.7.7.3"
types-pillow = "^9.3.0.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
